Simplify a lot and remove watchers
Also actually remove the git system command and just use the markers in
the file that git uses. I think it is quite fast to just parse the file
with string.match for the markers. Also, it can just as well work to go
to next and choose even if it isn't an "actual" git conflict.